Add basic components and routing to Workout App
Created main components of the app like the header, footer, card, and views for exercise list, create workout and user profile. Along with this, added the necessary routing for these components and updated the .gitignore file to include .env files. The main app was updated to use the router, components and imported styles. A basic fetch logic was also added to fetch exercises from a third-party API. Also updated 'package-lock.json' and 'index.html'.