Display exercise count in ExerciseList component
The number of exercises obtained from the API is now counted and the value is displayed in the ExerciseList component's header. This is achieved by including an 'exercisesCounter' variable in the 'useExercises' function. The exercises list is filtered server side based on the category selected, returning a count specific to the filtered list.