Refactor and implement workout component and related changes
This commit introduces a significant refactor by replacing several components: 'Session', 'UserProfile', 'UserSession', and 'Workout' with a new 'WorkoutList' component. The 'Exercise' import in 'Exercise.vue' and 'Exercise.ts' were updated to remove the '.ts' extension. In the router configuration, changed the URL routing for the 'Workouts' path to the 'WorkoutList' component, and the 'WorkoutCreate' route was updated. Server-side handling for workouts was added in 'filebase.js' and 'workouts.js' with updates to the data.json file to include workout data. Refactored 'categories.ts' and 'exercises.ts' files to 'use-categories.ts' and 'use-exercises.ts' respectively, for uniformity in file naming.